feature: DragDrop replaces individual implementations
