+++ src/Components/InfoStrip/InfoStripContainer.js
@@ -1,11 +1,10 @@
 import React, { memo } from 'react'
 import styled from 'styled-components'
-import { Draggable, Droppable } from 'react-beautiful-dnd'
 
 import { AutoFillPromos } from '../Autofill'
-import { BlockWrapper } from '../styles/ContentBlocks'
 import InfoStripBlock from './InfoStripBlock'
 import { useAppDispatch, useAppState } from '../../state-context'
+import DragDrop from '../DragDrop'
 
 const PromoButton = styled.input`
   background: none;
@@ -22,46 +21,20 @@ function InfoStripContainer() {
   const { promoBlocks } = useAppState()
 
   return (
-    <Droppable droppableId="promos">
-      {provided => (
-        <BlockWrapper ref={provided.innerRef} {...provided.droppableProps}>
-          <section>
-            <h3>Info Strip</h3>
-            <div>
-              <AutoFillPromos />
-              <PromoButton
-                type="button"
-                value="+"
-                onClick={() => dispatch({ type: 'addPromoBlock' })}
-              />
-            </div>
-          </section>
-          {promoBlocks.map((block, index) => (
-            <Draggable
-              key={index}
-              draggableId={`promo${index + 1}`}
-              index={index}
-            >
-              {provided => (
-                <div
-                  key={index}
-                  {...provided.draggableProps}
-                  ref={provided.ref}
-                  {...provided.dragHandleProps}
-                >
-                  <InfoStripBlock
-                    provided={provided}
-                    block={block}
-                    index={index}
-                  />{' '}
-                </div>
-              )}
-            </Draggable>
-          ))}
-          {provided.placeholder}
-        </BlockWrapper>
-      )}
-    </Droppable>
+    <DragDrop id="promos" stateItems={promoBlocks}>
+      <section>
+        <h3>Info Strip</h3>
+        <div>
+          <AutoFillPromos />
+          <PromoButton
+            type="button"
+            value="+"
+            onClick={() => dispatch({ type: 'addPromoBlock' })}
+          />
+        </div>
+      </section>
+      {(block, index) => <InfoStripBlock block={block} index={index} />}
+    </DragDrop>
   )
 }
